Update data generation method
This commit makes a change to how data is gathered and generated.

First we have a look in 'data' for locally maintained data (applies to
notation and roles). If not available locally we fetch it from the Cold
API.

For each datapoint we generate both standard and "pretty" JSON files and
store it in the 'public' directory.

If any failures are encountered, we log appropriately and fail the
build.

Local copies of Cold API data have been removed as part of this, to
ensure we aren't using stale (or in the case of API failures, very old)
data in our site build.

A move to httpx has been made, to allow for API basepath setting, and a
future move to async requests.